add ability to override s3 params key on upload
Useful when paths like site/service would have mapped on other systems to something like service.html. Default routing rules on s3 map it instead to index.html. With the ability to override the key, we can set the key to the file name for any of the html files to just be the file part (i.e. service) and it will return that file as html (since the content property type is still html).